fix: address review comments — Redis perf, type safety, CI fixes
- Batch _prune_old into single pipeline instead of per-conversation deletes
- Fold exists+get into pipeline in append_message to reduce round-trips
- Narrow exception handling from Exception to redis.RedisError
- Guard list_conversations against limit<=0 (consistent with JSONL backend)
- Fix _coord_ids type: NDArray[np.int64] (stores token IDs, not floats)
- Fix docstrings: RedisConfig Railway fallback claim, key layout "hash"→"string"
- Fix test_consciousness: TackingController kwarg, mypy None assertions, imports

+++ kernel/coordizer_v2/resonance_bank.py
@@ -63,7 +63,7 @@ class ResonanceBank:
         self.basin_mass: dict[int, float] = {}
         self.activation_counts: dict[int, int] = {}
         self._coord_matrix: NDArray[np.float64] | None = None
-        self._coord_ids: NDArray[np.float64] | None = None
+        self._coord_ids: NDArray[np.int64] | None = None
         self._dirty: bool = True
         self._domain_biases: list[DomainBias] = []
